simplify request params/response behaviours, reducing generated code
Similar to events, we can just implement the behaviour on the Basic
struct, and embedding that guy in each specific Request will satisfy
their interfaces too.

Furthermore, we now expose a `Name()` function that allows us to set the
RequestType ourselves.

Both of these changes reduce the amount of generated code.