use pointers to bools (#29)
* use pointers to bools

We need to be able to tell the difference between a false value, and an
unset value. Even though this change might make the UX kinda ugly if one
were to set the bools directly (e.g. `&[]bool{true}[0]`), it does make
the library more robust against future changes to the protocol.

* generate

* wow thank goodness for tests

Co-authored-by: Andrey Kaipov <andreykaipov@users.noreply.github.com>